[ALPS05153008] GPU:Fix build error
Details:
change 3747985 has build error for 32 bit project, success for 64 bit
project .

This may need revert when ARMs response

MTK-Commit-Id: 97cc3e62e11396dd944fd662dd75cca2660bfd76

CR-Id: ALPS05153008
Feature: OpenGL|ES

+++ drivers/misc/mediatek/gpu/gpu_mali/mali_bifrost/mali-r25p0/drivers/gpu/arm/midgard/backend/gpu/mali_kbase_clk_rate_trace_mgr.c
@@ -265,8 +265,8 @@ void kbase_clk_rate_trace_manager_notify_all(
 
 	/* Raise standard `power/gpu_frequency` ftrace event */
 	{
-		unsigned long new_rate_khz = new_rate;
-
+		//unsigned long new_rate_khz = new_rate;
+        uint64_t new_rate_khz = (uint64_t)new_rate;
 		do_div(new_rate_khz, 1000);
 		trace_gpu_frequency(new_rate_khz, clk_index);
 	}
